% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%||\%}
\alias{\%||\%}
\title{Reproduce functionality of JS || operator.}
\usage{
a \%||\% b
}
\arguments{
\item{a}{any R object}

\item{b}{R object to be returned if a does not exists or is null}
}
\value{
if a exists and is not null, return a, otherwise b.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

