% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listAllFun}
\alias{listAllFun}
\title{List all (public and private) functions within the specified package}
\usage{
listAllFun(pkg = "data.table")
}
\arguments{
\item{pkg}{character vector with the name of the package}
}
\value{
character string with names of functions within the package.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

