% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shift}
\alias{shift}
\title{Lag a time series within a panel}
\usage{
shift(x, lag = -1, dif = FALSE, relative = FALSE)
}
\arguments{
\item{x}{numeric vector to be shifted}

\item{dif}{}

\item{relative}{}

\item{shift.by}{number of legs}
}
\value{
a vector
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

