% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shiftData}
\alias{shiftData}
\title{Lead/lag/dif operations on a data.table}
\usage{
shiftData(data, var, replace = FALSE, by = NULL, lag = -1, dif = FALSE,
  relative = FALSE)
}
\arguments{
\item{data}{input dataset. Must be a data.table.}

\item{var}{character vector with names of variables that should be modified}

\item{replace}{replace the original variables. If FALSE (by default), the
modified variables are added to the original dataset, with a suffix added
to the name of the original variables.}

\item{by}{character vector with name(s) of by groups.}

\item{lag}{number of lags (negative) or leads (positive).}

\item{dif}{take a difference? Default is FALSE.}

\item{relative}{relative differences? Applicable only if dif=TRUE.}
}
\value{
data.table with the resulting dataset.
}
\description{
shiftData implements lead/lag/dif operations on a data.table and returns a
resulting dataset.
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

