% Generated by roxygen2 (4.0.2): do not edit by hand
\name{winsorize}
\alias{winsorize}
\title{Winsorize a vector}
\usage{
winsorize(x, method = c("IQR", "PERC"), k = 2, fraction = 0.01,
  na.rm = TRUE, trim = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{method}{method of winsorization.}

\item{k}{only applicable if method=="IQR". Specifies a number of
interquartile ranges from each quartile that define the winsorization limits.}

\item{fraction}{only applicable if method=="PERC"}

\item{na.rm}{omit missing values? Default is TRUE.}

\item{trim}{Trim extreme values rather setting them to a limit values?
Default is FALSE}
}
\value{
winsorized numeric vector
}
\description{
This function winsorizes an input numeric vector
}
\details{
.. TODO ..
}
\author{
Janko Cizel
}

